QA-Learning-Platform-2.0/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pydantic
httpx
